requirements.txt
Flask==3.0.0
gunicorn==21.2.0
tavily-python==0.2.8